+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - Реализованы базовые команды для чисел: `ZADD`, `ZSCORE`, `ZCARD`, `ZREM`, `ZRANGE`, `ZREVRANGE`.
   - Добавил тесты для методов: `ZADD`, `ZSCORE`, `ZCARD`, `ZREM`, `ZRANGE`, `ZREVRANGE`.
   - Реализованы базовые команды для чисел: `LPUSH`, `RPUSH`, `LPOP`, `RPOP`, `LLEN`, `LRANGE`.
+  - Добавил тесты для методов: `LPUSH`, `RPUSH`, `LPOP`, `RPOP`, `LLEN`, `LRANGE`.
 
 ## [0.9.2] - 2025-04-12